removing detected hack language
some php files were detected as hack files because those files contained only html code so added three lines of php code in order to detect those files as php only. Since Hack is not a famous language so removed it cause otherwise clients may get confuse. Reason for hack language instead of php on github-Hack files can also use the .php extension. Since Linguist considers each file independently and that file only contains HTML code, it's going to be difficult (impossible?) for Linguist to understand that it is not a Hack file.